Crée un composant CarouselTrack
