+++ src/custom/simplifions/components/CarouselTrack.vue
@@ -1,0 +1,96 @@
+<template>
+  <ul ref="track" class="carousel__track" @scroll.passive="onScroll">
+    <slot />
+  </ul>
+  <div v-show="hasOverflow" class="fr-mt-2w fr-pb-2w fr-grid-row fr-grid-row--center">
+    <button
+      type="button"
+      class="fr-link fr-link--sm fr-link--icon-left fr-icon-arrow-left-s-line fr-mr-2w"
+      :disabled="!canPrev"
+      @click="prev"
+    >
+      Précédent
+    </button>
+    <button
+      type="button"
+      class="fr-link fr-link--sm fr-link--icon-right fr-icon-arrow-right-s-line"
+      :disabled="!canNext"
+      @click="next"
+    >
+      Suivant
+    </button>
+  </div>
+</template>
+
+<script setup lang="ts">
+const props = defineProps<{ onResize?: () => void }>()
+
+const track = ref<HTMLElement | null>(null)
+const scrollLeft = ref(0)
+const scrollMax = ref(0)
+
+const hasOverflow = computed(() => scrollMax.value > 0)
+const canPrev = computed(() => scrollLeft.value > 4)
+const canNext = computed(() => scrollLeft.value < scrollMax.value - 4)
+
+const updateScrollState = () => {
+  if (!track.value) return
+  scrollMax.value = track.value.scrollWidth - track.value.clientWidth
+  scrollLeft.value = track.value.scrollLeft
+}
+
+const onScroll = () => {
+  if (!track.value) return
+  scrollLeft.value = track.value.scrollLeft
+}
+
+const itemWidth = () => {
+  const item = track.value?.querySelector('.carousel__item') as HTMLElement | null
+  return item ? item.offsetWidth + 16 : 300
+}
+
+const prev = () => track.value?.scrollBy({ left: -itemWidth(), behavior: 'smooth' })
+const next = () => track.value?.scrollBy({ left: itemWidth(), behavior: 'smooth' })
+
+const ro = new ResizeObserver(() => {
+  updateScrollState()
+  props.onResize?.()
+})
+
+onMounted(async () => {
+  if (!track.value) return
+  await nextTick()
+  updateScrollState()
+  ro.observe(track.value)
+})
+
+onUnmounted(() => ro.disconnect())
+
+defineExpose({
+  getEl: () => track.value,
+  updateScrollState
+})
+</script>
+
+<style scoped>
+.carousel__track {
+  display: flex;
+  flex-wrap: nowrap;
+  align-items: stretch;
+  gap: 1rem;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+  overflow-x: auto;
+  scroll-snap-type: x mandatory;
+  scrollbar-width: none;
+}
+.carousel__track::-webkit-scrollbar {
+  display: none;
+}
+
+button:disabled {
+  opacity: 0.4;
+  cursor: default;
+}
+</style>
